refactor: 아카이브 테이블 구조 변경
