+++ src/main/java/com/baro/archive/exception/ArchiveExceptionType.java
@@ -1,14 +1,15 @@
-package com.baro.template.exception;
+package com.baro.archive.exception;
 
 import com.baro.common.exception.RequestExceptionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @RequiredArgsConstructor
-public enum TemplateExceptionType implements RequestExceptionType {
+public enum ArchiveExceptionType implements RequestExceptionType {
 
-    INVALID_CATEGORY("TE01", "존재하지 않는 카테고리입니다.", HttpStatus.BAD_REQUEST),
-    INVALID_TEMPLATE("TE02", "존재하지 않는 템플릿입니다.", HttpStatus.NOT_FOUND),
+    NOT_EXIST_ARCHIVE("AR01", "존재하지 않는 아카이브 입니다.", HttpStatus.NOT_FOUND),
+    ARCHIVED_TEMPLATE("AR02", "이미 저장한 템플릿입니다.", HttpStatus.BAD_REQUEST),
+    NOT_ARCHIVED_TEMPLATE("AR03", "저장하지 않은 템플릿입니다.", HttpStatus.BAD_REQUEST),
     ;
 
     private final String errorCode;
