+++ src/main/java/com/baro/memo/application/TemporalMemoService.java
@@ -4,8 +4,6 @@ package com.baro.memo.application;
 import com.baro.member.domain.Member;
 import com.baro.member.domain.MemberRepository;
 import com.baro.memo.application.dto.ApplyCorrectionCommand;
-import com.baro.memo.application.dto.ArchiveTemporalMemoCommand;
-import com.baro.memo.application.dto.ArchiveTemporalMemoResult;
 import com.baro.memo.application.dto.DeleteTemporalMemoCommand;
 import com.baro.memo.application.dto.FindTemporalMemoHistoriesQuery;
 import com.baro.memo.application.dto.FindTemporalMemoHistoriesResult;
@@ -13,15 +11,11 @@ import com.baro.memo.application.dto.FindTemporalMemoResult;
 import com.baro.memo.application.dto.SaveTemporalMemoCommand;
 import com.baro.memo.application.dto.SaveTemporalMemoResult;
 import com.baro.memo.application.dto.UpdateTemporalMemoCommand;
-import com.baro.memo.domain.Memo;
 import com.baro.memo.domain.MemoContent;
-import com.baro.memo.domain.MemoRepository;
 import com.baro.memo.domain.TemporalMemo;
 import com.baro.memo.domain.TemporalMemoRepository;
 import com.baro.memo.exception.TemporalMemoException;
 import com.baro.memo.exception.TemporalMemoExceptionType;
-import com.baro.memofolder.domain.MemoFolder;
-import com.baro.memofolder.domain.MemoFolderRepository;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Comparator;
@@ -37,10 +31,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class TemporalMemoService {
 
-    private final MemoRepository memoRepository;
     private final TemporalMemoRepository temporalMemoRepository;
     private final MemberRepository memberRepository;
-    private final MemoFolderRepository memoFolderRepository;
 
     public SaveTemporalMemoResult saveTemporalMemo(SaveTemporalMemoCommand command) {
         Member member = memberRepository.getById(command.memberId());
@@ -52,18 +44,6 @@ public class TemporalMemoService {
         TemporalMemo temporalMemo = temporalMemoRepository.getById(command.temporalMemoId());
         temporalMemo.matchOwner(command.memberId());
         temporalMemo.updateContent(MemoContent.from(command.content()));
-    }
-
-    public ArchiveTemporalMemoResult archiveTemporalMemo(ArchiveTemporalMemoCommand command) {
-        Member member = memberRepository.getById(command.memberId());
-        TemporalMemo temporalMemo = temporalMemoRepository.getById(command.temporalMemoId());
-        temporalMemo.matchOwner(member.getId());
-        MemoFolder memoFolder = memoFolderRepository.getById(command.memoFolderId());
-        memoFolder.matchOwner(member.getId());
-
-        Memo memo = memoRepository.save(Memo.of(member, memoFolder, temporalMemo.getArchivingContent()));
-        temporalMemo.archivedAsMemo(memo);
-        return ArchiveTemporalMemoResult.from(memo);
     }
 
     public void deleteTemporalMemo(DeleteTemporalMemoCommand command) {
