+++ src/main/java/com/baro/memo/application/dto/ArchiveTemporalMemoResult.java
@@ -1,12 +1,12 @@
 package com.baro.memo.application.dto;
 
-import com.baro.memo.domain.Memo;
+import com.baro.archive.domain.Archive;
 
 public record ArchiveTemporalMemoResult(
         Long id
 ) {
 
-    public static ArchiveTemporalMemoResult from(Memo memo) {
-        return new ArchiveTemporalMemoResult(memo.getId());
+    public static ArchiveTemporalMemoResult from(Archive archive) {
+        return new ArchiveTemporalMemoResult(archive.getId());
     }
 }
