+++ src/main/java/com/baro/memo/presentation/TemporalMemoController.java
@@ -63,23 +63,6 @@ public class TemporalMemoController {
         return ResponseEntity.noContent().build();
     }
 
-    @PostMapping("/{temporalMemoId}/archive")
-    public ResponseEntity<Void> archiveAsMemo(
-            AuthMember authMember,
-            @RequestBody ArchiveTemporalMemoRequest request,
-            @PathVariable Long temporalMemoId
-    ) {
-        ArchiveTemporalMemoCommand command = new ArchiveTemporalMemoCommand(authMember.id(), temporalMemoId,
-                request.memoFolderId());
-        ArchiveTemporalMemoResult result = temporalMemoService.archiveTemporalMemo(command);
-
-        URI location = ServletUriComponentsBuilder.fromPath("/memos")
-                .path("/{id}")
-                .buildAndExpand(result.id())
-                .toUri();
-        return ResponseEntity.created(location).build();
-    }
-
     @DeleteMapping("/{temporalMemoId}")
     public ResponseEntity<Void> deleteTemporalMemo(
             AuthMember authMember,
