+++ src/main/java/com/baro/template/application/TemplateService.java
@@ -1,21 +1,10 @@
 package com.baro.template.application;
 
-import com.baro.member.domain.Member;
-import com.baro.member.domain.MemberRepository;
-import com.baro.memofolder.domain.MemoFolder;
-import com.baro.memofolder.domain.MemoFolderRepository;
-import com.baro.template.application.dto.ArchiveTemplateCommand;
-import com.baro.template.application.dto.ArchiveTemplateResult;
 import com.baro.template.application.dto.CopyTemplateCommand;
 import com.baro.template.application.dto.FindTemplateQuery;
 import com.baro.template.application.dto.FindTemplateResult;
-import com.baro.template.application.dto.UnArchiveTemplateCommand;
 import com.baro.template.domain.Template;
-import com.baro.template.domain.TemplateMember;
-import com.baro.template.domain.TemplateMemberRepository;
 import com.baro.template.domain.TemplateRepository;
-import com.baro.template.exception.TemplateException;
-import com.baro.template.exception.TemplateExceptionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
@@ -28,9 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class TemplateService {
 
     private final TemplateRepository templateRepository;
-    private final MemberRepository memberRepository;
-    private final MemoFolderRepository memoFolderRepository;
-    private final TemplateMemberRepository templateMemberRepository;
 
     @Transactional(readOnly = true)
     public Slice<FindTemplateResult> findTemplates(FindTemplateQuery query) {
@@ -40,36 +26,8 @@ public class TemplateService {
                 .map(FindTemplateResult::from);
     }
 
-    public ArchiveTemplateResult archiveTemplate(ArchiveTemplateCommand command) {
-        Member member = memberRepository.getById(command.memberId());
-        Template template = templateRepository.getById(command.templateId());
-
-        if (templateMemberRepository.existsByMemberIdAndTemplateId(member.getId(), template.getId())) {
-            throw new TemplateException(TemplateExceptionType.ARCHIVED_TEMPLATE);
-        }
-
-        MemoFolder folder = memoFolderRepository.getById(command.memoFolderId());
-        folder.matchOwner(member.getId());
-
-        TemplateMember templateMember = templateMemberRepository.save(TemplateMember.of(member, folder, template));
-        template.increaseSavedCount();
-
-        return ArchiveTemplateResult.from(templateMember);
-    }
-
     public void copyTemplate(CopyTemplateCommand command) {
         Template template = templateRepository.getById(command.templateId());
         template.increaseCopiedCount();
-    }
-
-    public void unarchiveTemplate(UnArchiveTemplateCommand command) {
-        Member member = memberRepository.getById(command.memberId());
-        Template template = templateRepository.getById(command.templateId());
-
-        TemplateMember templateMember = templateMemberRepository.getByMemberIdAndTemplateId(
-                member.getId(), template.getId());
-
-        templateMemberRepository.delete(templateMember);
-        template.decreaseSavedCount();
     }
 }
