+++ src/main/java/com/baro/template/application/dto/ArchiveTemplateResult.java
@@ -1,12 +1,12 @@
 package com.baro.template.application.dto;
 
-import com.baro.template.domain.TemplateMember;
+import com.baro.archive.domain.Archive;
 
 public record ArchiveTemplateResult(
         Long id
 ) {
 
-    public static ArchiveTemplateResult from(TemplateMember templateMember) {
-        return new ArchiveTemplateResult(templateMember.getId());
+    public static ArchiveTemplateResult from(Archive archive) {
+        return new ArchiveTemplateResult(archive.getId());
     }
 }
