+++ src/test/java/com/baro/memo/application/TemporalMemoServiceTest.java
@@ -9,28 +9,19 @@ import com.baro.member.domain.MemberRepository;
 import com.baro.member.fake.FakeMemberRepository;
 import com.baro.member.fixture.MemberFixture;
 import com.baro.memo.application.dto.ApplyCorrectionCommand;
-import com.baro.memo.application.dto.ArchiveTemporalMemoCommand;
 import com.baro.memo.application.dto.DeleteTemporalMemoCommand;
 import com.baro.memo.application.dto.FindTemporalMemoHistoriesQuery;
 import com.baro.memo.application.dto.FindTemporalMemoHistoriesResult;
 import com.baro.memo.application.dto.SaveTemporalMemoCommand;
 import com.baro.memo.application.dto.UpdateTemporalMemoCommand;
-import com.baro.memo.domain.Memo;
 import com.baro.memo.domain.MemoContent;
-import com.baro.memo.domain.MemoRepository;
 import com.baro.memo.domain.TemporalMemo;
 import com.baro.memo.domain.TemporalMemoRepository;
 import com.baro.memo.exception.MemoException;
 import com.baro.memo.exception.MemoExceptionType;
 import com.baro.memo.exception.TemporalMemoException;
 import com.baro.memo.exception.TemporalMemoExceptionType;
-import com.baro.memo.fake.FakeMemoRepository;
 import com.baro.memo.fake.FakeTemporalMemoRepository;
-import com.baro.memofolder.domain.MemoFolder;
-import com.baro.memofolder.domain.MemoFolderRepository;
-import com.baro.memofolder.exception.MemoFolderException;
-import com.baro.memofolder.exception.MemoFolderExceptionType;
-import com.baro.memofolder.fake.FakeMemoFolderRepository;
 import java.time.LocalDate;
 import java.util.Comparator;
 import java.util.List;
@@ -46,17 +37,12 @@ class TemporalMemoServiceTest {
     private TemporalMemoService temporalMemoService;
     private MemberRepository memberRepository;
     private TemporalMemoRepository temporalMemoRepository;
-    private MemoFolderRepository memoFolderRepository;
-    private MemoRepository memoRepository;
 
     @BeforeEach
     void setUp() {
         memberRepository = new FakeMemberRepository();
         temporalMemoRepository = new FakeTemporalMemoRepository();
-        memoFolderRepository = new FakeMemoFolderRepository();
-        memoRepository = new FakeMemoRepository();
-        temporalMemoService = new TemporalMemoService(memoRepository, temporalMemoRepository, memberRepository,
-                memoFolderRepository);
+        temporalMemoService = new TemporalMemoService(temporalMemoRepository, memberRepository);
     }
 
     @Test
@@ -145,82 +131,6 @@ class TemporalMemoServiceTest {
                 .isInstanceOf(TemporalMemoException.class)
                 .extracting("exceptionType")
                 .isEqualTo(TemporalMemoExceptionType.NOT_EXIST_TEMPORAL_MEMO);
-    }
-
-    @Test
-    void 끄적이는_메모_아카이브() {
-        // given
-        Member member = memberRepository.save(MemberFixture.memberWithNickname("nickname1"));
-        TemporalMemo temporalMemo = temporalMemoRepository.save(TemporalMemo.of(member, "testContent"));
-        Long memoFolderId = memoFolderRepository.save(MemoFolder.defaultFolder(member)).getId();
-
-        ArchiveTemporalMemoCommand command = new ArchiveTemporalMemoCommand(member.getId(), temporalMemo.getId(),
-                memoFolderId);
-
-        // when
-        temporalMemoService.archiveTemporalMemo(command);
-
-        // then
-        List<Memo> allMemos = memoRepository.findAll();
-        Memo savedMemo = allMemos.get(0);
-        assertAll(
-                () -> assertThat(allMemos).hasSize(1),
-                () -> assertThat(savedMemo.getContent()).isEqualTo(temporalMemo.getArchivingContent()),
-                () -> assertThat(temporalMemo.getMemo()).isEqualTo(savedMemo)
-        );
-    }
-
-    @Test
-    void 다른_사람의_끄적이는_메모_아카이브시_예외_발생() {
-        // given
-        Member member = memberRepository.save(MemberFixture.memberWithNickname("nickname1"));
-        Member otherMember = memberRepository.save(MemberFixture.memberWithNickname("nickname2"));
-        TemporalMemo temporalMemo = temporalMemoRepository.save(TemporalMemo.of(member, "testContent"));
-        Long memoFolderId = memoFolderRepository.save(MemoFolder.defaultFolder(otherMember)).getId();
-
-        ArchiveTemporalMemoCommand command = new ArchiveTemporalMemoCommand(otherMember.getId(), temporalMemo.getId(),
-                memoFolderId);
-
-        // when & then
-        assertThatThrownBy(() -> temporalMemoService.archiveTemporalMemo(command))
-                .isInstanceOf(TemporalMemoException.class)
-                .extracting("exceptionType")
-                .isEqualTo(TemporalMemoExceptionType.NOT_MATCH_OWNER);
-    }
-
-    @Test
-    void 존재하지_않는_메모_폴더에_아카이브시_예외_발생() {
-        // given
-        Member member = memberRepository.save(MemberFixture.memberWithNickname("nickname1"));
-        TemporalMemo temporalMemo = temporalMemoRepository.save(TemporalMemo.of(member, "testContent"));
-        Long notExistMemoFolderId = 1L;
-
-        ArchiveTemporalMemoCommand command = new ArchiveTemporalMemoCommand(member.getId(), temporalMemo.getId(),
-                notExistMemoFolderId);
-
-        // when & then
-        assertThatThrownBy(() -> temporalMemoService.archiveTemporalMemo(command))
-                .isInstanceOf(MemoFolderException.class)
-                .extracting("exceptionType")
-                .isEqualTo(MemoFolderExceptionType.NOT_EXIST_MEMO_FOLDER);
-    }
-
-    @Test
-    void 다른_사람의_메모_폴더에_아카이브시_예외_발생() {
-        // given
-        Member member = memberRepository.save(MemberFixture.memberWithNickname("nickname1"));
-        Member otherMember = memberRepository.save(MemberFixture.memberWithNickname("nickname2"));
-        TemporalMemo temporalMemo = temporalMemoRepository.save(TemporalMemo.of(otherMember, "testContent"));
-        Long memoFolderId = memoFolderRepository.save(MemoFolder.defaultFolder(member)).getId();
-
-        ArchiveTemporalMemoCommand command = new ArchiveTemporalMemoCommand(otherMember.getId(), temporalMemo.getId(),
-                memoFolderId);
-
-        // when & then
-        assertThatThrownBy(() -> temporalMemoService.archiveTemporalMemo(command))
-                .isInstanceOf(MemoFolderException.class)
-                .extracting("exceptionType")
-                .isEqualTo(MemoFolderExceptionType.NOT_MATCH_OWNER);
     }
 
     @Test
